moved router to App & created navbar component
