+++ frontend/src/routes/Search.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { Button, Grid, Paper } from '@mui/material/'
+
 import { FIELDS } from '../components/SubjectFields'
 import { SUBJECTS } from '../components/SubjectMethods'
